+++ resources/model_point_price_attributes.go
@@ -6,5 +6,5 @@ package resources
 
 type PointPriceAttributes struct {
 	// Amount of `urmo` tokens for one point
-	Urmo int32 `json:"urmo"`
+	Urmo int64 `json:"urmo"`
 }
